requirements.txt
python-telegram-bot==20.7
openpyxl==3.1.2
pandas==2.1.4
python-docx==1.1.0
